Refactoring TokenStore tests into a single parameterized test
